refactor(footer): show social/donation on homepage (#142)
* refactor: lint

* fix: linting

* refactor(footer): show social/donation on homepage

* fix: review suggestions

* refactor(technology-logos): use css grid
