--- src/components/UI/Donation.svelte
@@ -1,10 +1,0 @@
-<style lang="sass">
-    input[type=image]
-        outline: none
-</style>
-
-<form action="https://www.paypal.com/cgi-bin/webscr" method="post" target="_blank">
-    <input type="hidden" name="cmd" value="_s-xclick" />
-    <input type="hidden" name="hosted_button_id" value="YN3NWVKZJMT6C" />
-    <input type="image" src="https://www.paypalobjects.com/en_US/NL/i/btn/btn_donateCC_LG.gif" border="0" name="submit" title="PayPal - The safer, easier way to pay online!" alt="Donate with PayPal button" />
-</form>
