+++ src/components/UI/Layout/MainLayout.svelte
@@ -1,31 +1,26 @@
 <script>
-    import RunmeFooter from './RunmeFooter.svelte'
-    import JexiaFooter from './JexiaFooter.svelte'
-    import Header from './Header.svelte'
-    import TechnologyLogos from './TechnologyLogos.svelte'
+  import Header from './Header.svelte'
+  import MainFooter from '../Footer/MainFooter.svelte'
+  import SubFooter from '../Footer/SubFooter.svelte'
 
-    export let showRunmeFooter = true
-    export let showTechnologyIcons = false
-    export let title = ''
+  export let title = ''
+  export let showTechnologyIcons = false
 
-    const suffix = 'Runme.io'
-    const pageTitle = title ? `${title} - ${suffix}` : suffix
+  const suffix = 'Runme.io'
+  const pageTitle = title ? `${title} - ${suffix}` : suffix
 </script>
 
 <svelte:head>
-    <title>{pageTitle}</title>
+  <title>{pageTitle}</title>
 </svelte:head>
 
 <div class="page-container">
-    <Header/>
+  <Header/>
 
-    <main>
-        <div class="container">
-            <slot/>
-        </div>
-    </main>
+  <main class="container">
+    <slot/>
+  </main>
 
-    {#if showTechnologyIcons}<TechnologyLogos/>{/if}
-    {#if showRunmeFooter}<RunmeFooter/>{/if}
-    <JexiaFooter/>
+  <SubFooter {showTechnologyIcons}/>
+  <MainFooter/>
 </div>
