+++ src/routes/build.svelte
@@ -1,113 +1,118 @@
 <script>
-    import FixedHeader from '../components/UI/Layout/FixedHeader.svelte'
-    import JexiaFooter from '../components/UI/Layout/JexiaFooter.svelte'
-    import { goto } from '@sapper/app'
-    import { build } from '../components/Stores/Build'
-    import { header } from '../components/Stores/Header'
-    import { application } from '../components/Stores/Application'
-    import { queryParam, isBase64 } from '../Helpers'
-    import { onDestroy } from 'svelte'
-    import CliWindow from '../components/UI/Cli/CliWindow.svelte'
+  import FixedHeader from '../components/UI/Layout/FixedHeader.svelte'
+  import MainFooter from '../components/UI/Footer/MainFooter.svelte'
+  import { goto } from '@sapper/app'
+  import { build } from '../components/Stores/Build'
+  import { header } from '../components/Stores/Header'
+  import { application } from '../components/Stores/Application'
+  import { queryParam, isBase64 } from '../Helpers'
+  import { onDestroy } from 'svelte'
+  import CliWindow from '../components/UI/Cli/CliWindow.svelte'
 
-    const buildId = queryParam().get('build_id')
+  const buildId = queryParam().get('build_id')
 
-    let building = true
-    let buildLog = ''
-    let buildErrorMsg
-    let countFrom = 0
-    let workingOn = 'Build'
-    let cliTitle = ''
+  let building = true
+  let logContent = ''
+  let buildErrorMsg
+  let workingOn = 'Build'
+  let cliTitle = ''
 
-    const showBuildError = (error) => {
-        header.isFailed(true)
-        buildErrorMsg = error
-        building = false
+  const showBuildError = (error) => {
+    header.isFailed(true)
+    buildErrorMsg = error
+    building = false
+  }
+
+  const done = (buildId) => {
+    if (process.browser) {
+      goto(`/show?build_id=${buildId}`)
+    }
+  }
+
+  const collectLog = (buildLog, deployLog) => {
+    const log = []
+
+    if (buildLog) {
+      if (isBase64(buildLog)) {
+        buildLog = atob(buildLog)
+      }
+
+      log.push(buildLog)
     }
 
-    const done = (buildId) => {
-        if (process.browser) {
-            goto(`/show?build_id=${buildId}`)
-        }
+    if (deployLog) {
+      workingOn = 'Deployment'
+
+      if (isBase64(deployLog)) {
+        deployLog = atob(deployLog)
+      }
+
+      log.push(deployLog)
     }
 
-    const collectLog = (build_log, deploy_log) => {
-        let log = []
+    logContent = log.join('\r\n')
+  }
 
-        if (build_log) {
-            if (isBase64(build_log)) {
-                build_log = atob(build_log)
-            }
+  $: if (Object.keys($application).length) {
+    const { repo_branch: repoBranch, repo_name: repoName } = $application
+    cliTitle = `${repoName}:${repoBranch}`
+  }
 
-            log.push(build_log)
-        }
-
-        if (deploy_log) {
-            workingOn = 'Deployment'
-
-            if (isBase64(deploy_log)) {
-                deploy_log = atob(deploy_log)
-            }
-
-            log.push(deploy_log)
-        }
-
-        buildLog = log.join('\r\n')
+  const unsubscribe = build.subscribe(({
+    error,
+    status,
+    deploy_log: deployLog,
+    build_log: buildLog,
+    id,
+    app_id: appId,
+    created_at: createdAt,
+  }) => {
+    if (error) {
+      showBuildError(error.message)
+      return
     }
 
-    $: if (Object.keys($application).length) {
-        const { repo_branch, repo_name } = $application
-        cliTitle = `${repo_name}:${repo_branch}`
+    // fetch application info when the cliTitle has not been set with the repo info
+    if (!cliTitle) {
+      application.get(appId)
     }
 
-    const unsubscribe = build.subscribe(({ error, status, deploy_log, build_log, id, app_id, created_at }) => {
-        if (error) {
-            showBuildError(error.message)
-            return;
-        }
+    // update the header values
+    header.showCountUp(createdAt, 'Build time')
 
-        // fetch application info when the cliTitle has not been set with the repo info
-        if (!cliTitle) {
-            application.get(app_id)
-        }
-
-        // update the header values
-        header.showCountUp(created_at, 'Build time')
-
-        if (status === 'fail') {
-            showBuildError(`Build failed`)
-        }
-
-        if (status === 'done') {
-            done(id)
-        }
-
-        // collect the log and combine them
-        collectLog(build_log, deploy_log)
-    })
-
-    // no repo url? Redirect back
-    if (buildId === '') {
-        showBuildError('"build_id" is missing')
-    } else {
-        // start fetching the log
-        build.get(buildId, true)
+    if (status === 'fail') {
+      showBuildError('Build failed')
     }
 
-    onDestroy(unsubscribe)
+    if (status === 'done') {
+      done(id)
+    }
+
+    // collect the log and combine them
+    collectLog(buildLog, deployLog)
+  })
+
+  // no repo url? Redirect back
+  if (buildId === '') {
+    showBuildError('"build_id" is missing')
+  } else {
+    // start fetching the log
+    build.get(buildId, true)
+  }
+
+  onDestroy(unsubscribe)
 </script>
 
 <svelte:head>
-    <title>Runme.io - Building your application</title>
+  <title>Runme.io - Building your application</title>
 </svelte:head>
 
 <div class="page-container">
-    <FixedHeader/>
+  <FixedHeader/>
 
-    <div class="container">
-        <main>
-            <CliWindow {workingOn} working={building} error={buildErrorMsg} log={buildLog} title={cliTitle}/>
-        </main>
-    </div>
+  <main class="container">
+    <CliWindow {workingOn} working={building} error={buildErrorMsg} log={logContent} title={cliTitle}/>
+  </main>
 
-    <JexiaFooter/>
+  <MainFooter/>
 </div>
